Refactor translations module
It was an attempt to get the proper translations paths, but due to Vite
shortcomings, that differently than Node.js `import()`, doesn't accepts
functions nor variables to define the import path.

So the template string must begin with "./" or "../" for Vite to
correctly make the import.

This modifies the translation module to become a locale filter to
provide only the locales we support on our application.